backend/requirements.txt
# MedVision AI - Backend Dependencies
# Compatible with Python 3.9 - 3.13

# Web Framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
httpx>=0.25.0,<1.0.0

# Data Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Authentication
PyJWT>=2.8.0,<3.0.0
bcrypt>=4.0.0,<5.0.0

# Database (local SQLite)
sqlalchemy>=2.0.0,<3.0.0

# Document Processing
pymupdf>=1.23.0,<2.0.0
pillow>=10.0.0,<12.0.0

# Environment & AI
python-dotenv>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0

# Firebase (Firestore Database)
firebase-admin>=6.0.0,<7.0.0
